# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Reckon API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /profile/resume:
    put:
      operationId: uploadResume
      tags: [profile]
      summary: Upload or update resume
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                resume:
                  type: string
                  format: binary
                resume_text:
                  type: string
      responses:
        "200":
          description: Resume uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeUploadResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        hint:
          type: string
      required:
        - error
    Profile:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
          nullable: true
        resume_text:
          type: string
          nullable: true
        resume_url:
          type: string
          nullable: true
        subscription_type:
          type: string
          enum: [free, payg, monthly]
        jobs_count:
          type: integer
        amount_owed:
          type: number
        created_at:
          type: string
        updated_at:
          type: string
          nullable: true
      required:
        - id
        - subscription_type
    ProfileResponse:
      type: object
      properties:
        profile:
          $ref: "#/components/schemas/Profile"
      required:
        - profile
    UpdateProfileBody:
      type: object
      properties:
        full_name:
          type: string
    ResumeUploadResponse:
      type: object
      properties:
        profile:
          $ref: "#/components/schemas/Profile"
        resume_text:
          type: string
      required:
        - profile
...